separate type tests into `expression.test-d.ts`
